backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
polars==0.19.12
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
chardet==5.2.0
openpyxl==3.1.2
requests==2.31.0
supabase==2.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9